Support moveing VF which is eth0 on the host to the namespace
When using multus pluing the first interface will be eth0 and
the others will be netX. We need to make sure that before
we move VF which is eth0 to the container namespace it should
first be renamed to temp name before moving to the namespace.
This is to avoid conflict with the eth0 interface which already
exist in the namespace.

issue: 1520354
Change-Id: Ieecb8a7ca64f9c982ea6f204cd7a004bb9d2f0b8
Signed-off-by: mamduhala <mamduhala@mellanox.com>